+++ src/_components/PieDePagina/Footer.js
@@ -29,13 +29,12 @@ export default function AutoGrid() {
       <Grid wrap="nowrap" container spacing={1}>
         <Grid item xs={7}>
           <h5 className="colortext MuiTypography-root makeStyles-title-3 MuiTypography-h6" >El Diquecito Grupo 3</h5>
-          <Typography className="colortext">{message}</Typography>    
+          <Typography className="colortext">{message}</Typography>
         </Grid>
         <Grid item xs={5}>
           <h5 className="colortext MuiTypography-root makeStyles-title-3 MuiTypography-h6">Links</h5>
                   <li className="colortext MuiTypography-root makeStyles-title-3 MuiTypography-h6"><a className="colortext" href="https://es-la.facebook.com/diquecito.nuevocomplejo/">Facebook</a></li>
                   <li className="colortext MuiTypography-root makeStyles-title-3 MuiTypography-h6"><a className="colortext" href="https://www.instagram.com/explore/locations/1001217278/nuevo-complejo-diquecito">Instagram</a></li>
-              
         </Grid>
       </Grid>
       </Paper>
